split classes into cpp/hpp pairs
